add inheritance and abstract base class, modularize code, create handlers
