+++ Makefile
@@ -12,8 +12,8 @@ UTILS_TEST=test/utils_test.cpp
 SERVER_TEST=test/server_test.cpp
 PARSER_TEST=test/config_parser_test.cc
 
-all: server.o session.o main.o config_parser.o utils.o HttpRequest.o http_response.o
-	g++ -o web-server main.o server.o session.o config_parser.o utils.o HttpRequest.o http_response.o  $(LDFLAGS) $(CXXFLAGS) 
+all: server.o session.o main.o config_parser.o utils.o HttpRequest.o http_response.o handler.o
+	g++ -o web-server main.o server.o session.o config_parser.o utils.o HttpRequest.o http_response.o handler.o $(LDFLAGS) $(CXXFLAGS) 
 
 server.o: server.cpp server.h
 	g++ -c server.cpp $(LDFLAGS) $(CXXFLAGS)
@@ -39,7 +39,6 @@ http_response.o: http_response.h http_response.cpp
 handler.o: handler.h handler.cpp
 	g++ -c handler.cpp $(LDFLAGS) $(CXXFLAGS)
 
-
 .PHONY: clean, all, test
 
 test: 
@@ -49,7 +48,7 @@ test:
 	g++ -std=c++0x -isystem ${GTEST_DIR}/include -I${GTEST_DIR} -pthread -c ${GTEST_DIR}/src/gtest-all.cc
 	ar -rv libgtest.a gtest-all.o
 	# Build tests
-	g++ -isystem ${GTEST_DIR}/include ${SESSION_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o HttpRequest.o http_response.o -o session_test ${LDFLAGS} ${CXXFLAGS}
+	g++ -isystem ${GTEST_DIR}/include ${SESSION_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o HttpRequest.o http_response.o handler.o -o session_test ${LDFLAGS} ${CXXFLAGS}
 	g++ -isystem ${GTEST_DIR}/include ${UTILS_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o -o utils_test ${LDFLAGS} ${CXXFLAGS}
 	g++ -isystem ${GTEST_DIR}/include ${PARSER_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o -o config_parser_test ${LDFLAGS} ${CXXFLAGS}
 	g++ -isystem ${GTEST_DIR}/include ${SERVER_TEST} ${GTEST_DIR}/src/gtest_main.cc libgtest.a utils.o config_parser.o -o server_test ${LDFLAGS} ${CXXFLAGS}
